requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
SQLAlchemy>=2.0
pydantic>=2.6
python-multipart>=0.0.9
python-dotenv>=1.0
pdfminer.six>=20231228
python-docx>=1.1.0
email-validator>=2.0
psycopg2-binary
